requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.12
djangorestframework==3.8.2
idna==2.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
